requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# aquery/query.py: 5
# aquery/query_model.py: 2
mysql_connector_python == 8.0.17
